Push GradleDistribution implementation details to internal subclass
