+++ org.eclipse.buildship.core/src/main/java/org/eclipse/buildship/core/GradleDistribution.java
@@ -10,14 +10,10 @@ package org.eclipse.buildship.core;
 
 import java.io.File;
 import java.net.URI;
-import java.net.URISyntaxException;
 
 import org.gradle.tooling.GradleConnector;
 
-import com.google.common.base.Objects;
-import com.google.common.base.Optional;
-import com.google.common.base.Preconditions;
-
+import org.eclipse.buildship.core.internal.DefaultGradleDistribution;
 import org.eclipse.buildship.core.util.gradle.GradleDistributionInfo;
 
 /**
@@ -26,91 +22,22 @@ import org.eclipse.buildship.core.util.gradle.GradleDistributionInfo;
  *
  * @author Etienne Studer
  */
-public class GradleDistribution {
+public abstract class GradleDistribution {
 
-    private final GradleDistributionInfo distributionInfo;
+    public abstract GradleDistributionInfo getDistributionInfo();
 
-    private GradleDistribution(GradleDistributionType type, String configuration) {
-        this(new GradleDistributionInfo(type, configuration));
-    }
+    public abstract GradleDistributionType getType();
 
-    private GradleDistribution(GradleDistributionInfo distributionInfo) {
-        Optional<String> validationError = distributionInfo.validate();
-        Preconditions.checkArgument(!validationError.isPresent(), validationError.or(""));
-        this.distributionInfo = distributionInfo;
-    }
-
-    public GradleDistributionInfo getDistributionInfo() {
-        return this.distributionInfo;
-    }
-
-    public GradleDistributionType getType() {
-        return this.distributionInfo.getType();
-    }
-
-    public String getConfiguration() {
-        return this.distributionInfo.getConfiguration();
-    }
+    public abstract String getConfiguration();
 
     /**
      * Configures the specified connector with this distribution.
      *
      * @param connector the connector to configure
      */
-    public void apply(GradleConnector connector) {
-        switch (this.distributionInfo.getType()) {
-            case LOCAL_INSTALLATION:
-                connector.useInstallation(new File(this.distributionInfo.getConfiguration()));
-                break;
-            case REMOTE_DISTRIBUTION:
-                connector.useDistribution(createURI(this.distributionInfo.getConfiguration()));
-                break;
-            case VERSION:
-                connector.useGradleVersion(this.distributionInfo.getConfiguration());
-                break;
-            case WRAPPER:
-                connector.useBuildDistribution();
-                break;
-            default:
-                throw new GradlePluginsRuntimeException("Invalid distribution type: " + this.distributionInfo.getType());
-        }
-    }
+    public abstract void apply(GradleConnector connector);
 
-    public String serializeToString() {
-        return this.distributionInfo.serializeToString();
-    }
-
-    private static URI createURI(String path) {
-        try {
-            return new URI(path);
-        } catch (URISyntaxException e) {
-            throw new RuntimeException(e.getMessage());
-        }
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (this == other) {
-            return true;
-        }
-
-        if (other == null || getClass() != other.getClass()) {
-            return false;
-        }
-
-        GradleDistribution that = (GradleDistribution) other;
-        return Objects.equal(this.distributionInfo, that.distributionInfo);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(this.distributionInfo, this.distributionInfo);
-    }
-
-    @Override
-    public String toString() {
-        return this.distributionInfo.toString();
-    }
+    public abstract String serializeToString();
 
     /**
      * Creates a reference to a local Gradle installation.
@@ -120,7 +47,7 @@ public class GradleDistribution {
      * @see org.gradle.tooling.GradleConnector#useInstallation(java.io.File)
      */
     public static GradleDistribution forLocalInstallation(String installationDir) {
-        return new GradleDistribution(GradleDistributionType.LOCAL_INSTALLATION, installationDir);
+        return DefaultGradleDistribution.forLocalInstallation(installationDir);
     }
 
     /**
@@ -131,7 +58,7 @@ public class GradleDistribution {
      * @see org.gradle.tooling.GradleConnector#useInstallation(java.io.File)
      */
     public static GradleDistribution forLocalInstallation(File installationDir) {
-        return new GradleDistribution(GradleDistributionType.LOCAL_INSTALLATION, installationDir.getAbsolutePath());
+        return DefaultGradleDistribution.forLocalInstallation(installationDir);
     }
 
     /**
@@ -143,7 +70,7 @@ public class GradleDistribution {
      * @see org.gradle.tooling.GradleConnector#useDistribution(java.net.URI)
      */
     public static GradleDistribution forRemoteDistribution(String distributionUri) {
-        return new GradleDistribution(GradleDistributionType.REMOTE_DISTRIBUTION, distributionUri);
+        return DefaultGradleDistribution.forRemoteDistribution(distributionUri);
     }
 
     /**
@@ -155,7 +82,7 @@ public class GradleDistribution {
      * @see org.gradle.tooling.GradleConnector#useDistribution(java.net.URI)
      */
     public static GradleDistribution forRemoteDistribution(URI distributionUri) {
-        return new GradleDistribution(GradleDistributionType.REMOTE_DISTRIBUTION, distributionUri.toString());
+        return DefaultGradleDistribution.forRemoteDistribution(distributionUri);
     }
 
     /**
@@ -167,7 +94,7 @@ public class GradleDistribution {
      * @see org.gradle.tooling.GradleConnector#useGradleVersion(String)
      */
     public static GradleDistribution forVersion(String version) {
-        return new GradleDistribution(GradleDistributionType.VERSION, version);
+        return DefaultGradleDistribution.forVersion(version);
     }
 
     /**
@@ -177,10 +104,10 @@ public class GradleDistribution {
      * @see org.gradle.tooling.GradleConnector#useBuildDistribution()
      */
     public static GradleDistribution fromBuild() {
-        return new GradleDistribution(GradleDistributionType.WRAPPER, null);
+        return DefaultGradleDistribution.fromBuild();
     }
 
     public static GradleDistribution fromDistributionInfo(GradleDistributionInfo distributionInfo) {
-        return new GradleDistribution(distributionInfo);
+        return DefaultGradleDistribution.fromDistributionInfo(distributionInfo);
     }
 }
